// src/common/exceptions/custom.exception.codes.enum.ts
export enum ExceptionCodes {
  E_INTERNAL_SERVER_ERROR = 'E_INTERNAL_SERVER_ERROR',
  E_INVALID_CREDENTIALS = 'E_INVALID_CREDENTIALS',
  E_MUST_BE_A_STRING = 'E_MUST_BE_A_STRING',
  E_MUST_BE_A_NUMBER = 'E_MUST_BE_A_NUMBER',
  E_INVALID_EMAIL = 'E_INVALID_EMAIL',
  E_INVALID_MIN_LENGTH = 'E_INVALID_MIN_LENGTH',
  E_INVALID_ACCESS_LEVEL = 'E_INVALID_ACCESS_LEVEL',
  E_NOT_FOUND = 'E_NOT_FOUND',
  E_INVALID_AUTHORIZATION_HEADER = 'E_INVALID_AUTHORIZATION_HEADER',
  E_UNAUTHORIZED = 'E_UNAUTHORIZED',
}